requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-chroma>=0.1.2

# Vector Store & Embeddings
chromadb>=0.5.0

# LLM Integration (Ollama)
ollama>=0.4.0

# Document Processing
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
html2text>=2024.0.0
lxml>=5.0.0

# HTTP Client
httpx>=0.27.0

# CLI & Configuration
typer>=0.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Text Processing
regex>=2024.0.0

# Utilities
rich>=13.0.0  # Better terminal output
tenacity>=8.0.0  # Retry logic
